+++ examples/button/main.py
@@ -32,6 +32,8 @@ b0 = NexButton(nh, 0, 1, "b0")
 button_text = "btn txt"
 background_color_value = 63488  # red
 font_color_value = 31           # blue
+x_offset = 20
+y_offset = 20
 
 # request the text of button "b0"
 print('Requesting button "{}" text ...'.format(b0.name))
@@ -115,6 +117,26 @@ print()
 # sanity check
 if response != font_color_value:
     print('WARNING: GET value did not match SET value')
+
+time.sleep(1)
+
+# request the x/y position of button "b0" again
+print('Requesting x/y position of button "{}" ...'.format(b0.name))
+x_position = b0.Get_place_xcen()
+y_position = b0.Get_place_ycen()
+print('Position of button "{}" is: "x={}", "y={}"'.
+      format(b0.name, x_position, y_position))
+print()
+
+x_position += x_offset
+y_position += y_offset
+
+# modify the x/y position of button "b0"
+print('Set x/y position of button "{}" to "x={}", "y={}"'.
+      format(b0.name, x_position, y_position))
+b0.Set_place_xcen(x_position)
+b0.Set_place_ycen(y_position)
+print()
 
 print('Returning to REPL in 5 seconds')